Resolve issue: value too great for base
Following the comment part of the script, "time" needs to be in fomat HH:mm. Therefore there is a chance to have leading "0" which is interpreted as octal value.
For example this can happen when value for HH is set to a value 07 < x <10. The interpreter assumes to have a octal value then leading to wrong calculation for rtcwake.
Fix forced to interpret values with leading "0" to base 10.